suraksheet-extractor/requirements.txt
pika==1.2.0
minio==7.1.0
pytesseract==0.3.9
Pillow==8.3.2
requests==2.26.0
opencv-python-headless==4.7.0.68
psycopg2-binary==2.9.3  
numpy==1.23.5  
cryptography==36.0.2  
elasticsearch==8.11.1
